+++ src/manager.rs
@@ -69,6 +69,59 @@ fn pipe_closed(_pipe: Pipe, event: PipeEvent) {
     }
 }
 
+/// Merge per-client-statistics from client thread into global parent statistics
+fn merge_from_worker(
+    parent_request: &GooseRequest,
+    client_request: &GooseRequest,
+    config: &GooseConfiguration,
+) -> GooseRequest {
+    // Make a mutable copy where we can merge things
+    let mut merged_request = parent_request.clone();
+    // Iterate over client response times, and merge into global response time
+    merged_request.response_times = stats::merge_response_times(
+        merged_request.response_times,
+        client_request.response_times.clone(),
+    );
+    // Increment total response time counter.
+    merged_request.total_response_time += &client_request.total_response_time;
+    // Increment count of how many resposne counters we've seen.
+    merged_request.response_time_counter += &client_request.response_time_counter;
+    // If client had new fastest response time, update global fastest response time.
+    merged_request.min_response_time = stats::update_min_response_time(
+        merged_request.min_response_time,
+        client_request.min_response_time,
+    );
+    // If client had new slowest response time, update global slowest resposne time.
+    merged_request.max_response_time = stats::update_max_response_time(
+        merged_request.max_response_time,
+        client_request.max_response_time,
+    );
+    // Increment total success counter.
+    merged_request.success_count += &client_request.success_count;
+    // Increment total fail counter.
+    merged_request.fail_count += &client_request.fail_count;
+    // Only accrue overhead of merging status_code_counts if we're going to display the results
+    if config.status_codes {
+        for (status_code, count) in &client_request.status_code_counts {
+            let new_count;
+            // Add client count into global count
+            if let Some(existing_status_code_count) =
+                merged_request.status_code_counts.get(&status_code)
+            {
+                new_count = *existing_status_code_count + *count;
+            }
+            // No global count exists yet, so start with client count
+            else {
+                new_count = *count;
+            }
+            merged_request
+                .status_code_counts
+                .insert(*status_code, new_count);
+        }
+    }
+    merged_request
+}
+
 pub fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
     // Creates a TCP address.
     let address = format!(
@@ -212,7 +265,7 @@ pub fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
                                 if let Some(parent_request) =
                                     goose_attack.merged_requests.get(&request_key)
                                 {
-                                    merged_request = crate::merge_from_client(
+                                    merged_request = merge_from_worker(
                                         parent_request,
                                         &request,
                                         &goose_attack.configuration,